fix process control propagation
Some process control functions (kill,reset,throw_it)
can lead to exception throws, which potentially change
the set of child objects of a process.

When propagating the current operation to a process'
descendants, we therefore need a stable child_objects
vector _copy_, instead of a pointer.

For the other operations, this patch now uses a
reference for brevity.