reorganise common include structure
With this change, each "test set" (i.e. top-level directory in tests/)
has an accompanying include directory in include/<set>, which is passed
to the compiler as an additional include path for each test in the
current set.

This avoids hardcoding each and every shared include directory for
all tests in verify.pl.