Fix configuration and compilation on Mac OS X
Previously, there were only TARGET_ARCHs macosx for ppc and macosx386 for
i386. In recent revisions, the macosx386 TARGET_ARCH was actually
configured for x86_64, which has primarily an impact on the configuration
of QuickThreads. Since the architecture was not passed with the "-arg"
option to the compiler, it could happen that there would be a mismatch
between the default compiled C++ files and the compiled Assembler files.
This causes the build to fail.

Now, the architecture ppc, i386, and x86_64 is explicitly passed to the
C/C++ compilers and assembler ensuring a consistent build. This enables
additionally to easily cross-compile SystemC for the three supported
architectures of Mac OS X on the same machine.

Signed-off-by: Philipp A. Hartmann <philipp.hartmann@offis.de>