Ensure correct sign encoding of calculated 9 bit integer reference value despite compiler optimization in test systemc/datatypes/int/arith/arith01/arith01.cpp.
The above test routinely fails with an assertion error on Linux platforms if compiled with optimization -O2 or higher (g++). The test increments/decrements a 9 bit sc_signed variable and compares it with the expected result calculated in a 9 bit signed integer stored in a bit field struct. However, the size of the integer member gets neglected for the calculation and comparison with turned on optimization. In consequence, the integer member variable doesn't anymore encode correctly the sign over its whole siye in case of an underflow or overflow within its 9 bit bounds. The problem is fixed by applying a modulo 0x200 operation to the integer member of the int9 struct.

Additionally, the output statements are rearranged to always output the current value of the y and v.q before their equality is checked with assert().