Turn of IEEE Std 1666 deprecation warning for sc_bit in systemc/misc/stars/star116072/test.cpp
Depending on the platform and chosen optimization level, the sc_bit deprecation warning gets mingled with the surrounding output, as sc_bit is instantiated for the first time within a cout statement. This causes the regression test to fail with a diff error (e.g., on OS X 10.7.4 Lion (x84_64) with g++-4.2.1).