Detect more reliably 64 bit architecture on Mac OS X
Until Mac OS X Snow Leopard (<= 10.6.8), even 64 bit capable platforms are reported by the kernel to be 32 bit, i.e., ppc instead of ppc64 and i386 instead of x86_64. The command `sysctl -n hw.cpu64bit_capable` outputs 1 if the CPU support 64 bit and otherwise 0. This command is used in the OS-X-specific section of verify.pl. If sysctl returns 0 or another unexpected value (e.g., due to the potential non-existance of the key in older OS X versions), the processor reported by the kernel is kept as default.