Fixed incorrect GCC i386/x86_64 function call alignment
Since GCC-3.x, the i386/x86_64 GCC internal ABI requires the stack to be
aligned on 16-byte boundary at the 'call' instruction time. This is required
by GCC to generate code able to correctly align memory areas in the stack
that need to be accessed with 16-byte aligned addresses (namely by SSE
instructions). if the area is not aligned, then a SIGSEGV occurs at the
SSE instruction execution. The qt stack switch lay-out has been slightly
modified to accomodate with this requirement. In addition, the x86_64-dependent
switch code has also been reduced by 25% with the removal of unnecessary
scratch register savings.