fix stack alignment assumptions on non-QuickThread platforms (#39)
In addition to the fixes in qt/md/{i386,iX86_64}.[hs] contributed by STM,
this patch adds an attribute to the SC_(C)THREAD entry functions to force
the 16-byte alignment on further platforms.

This includes especially MinGW32 and Cygwin32 platforms (which don't use
QuickThreads) and may be sufficient for other non-Intel architectures as
well.