simple_target_socket: fixup static senstitivity in nb2b_thread (#13, part 1)
Quoting from #13:
  The nb2b_thread dynamic process in the simple_target_socket is created with
  a static sensitivity list but that event gets deleted in the nb2b_thread
  implementation. The standard doesn't say anything about changing the static
  sensitivity of a spawned process during simulation but it seems that the
  intention is that this should not be allowed. [...]

This patch drops the dynamic event allocation related to the nb2b_thread process
and adds an explicit event to the internal `process_handle_class`, to which the
process is then sensitive.

Secondly, the `m_wakeup` event is then no longer needed, as the process handle
information is updated when selecting the next free process handle.

Last, but not least, the process handle objects are freed up from the destructor
of the process_handle_list class.