Fixed missing initialization and checking of member variables
Member variables of callback_binder_fw and callback_binder_bw
are not initialized. This leads to crashes when the corresponding
TLM-2 functions (b_transport(), etc.) are called while the callbacks
have not been set (set_callbacks()).

In addition, in the implementation of TLM-2 functions, the callback
objects are tested for emptiness but their pointer may well be 0
which also causes a crash at execution time.