Fixed dummy interface used for export binding
Because of a fundamental limitation of sc_export, which cannot
have multiple interfaces bound to it (contrary to sc_port),
the actual implementation of multi_passthrough_target_socket
(and probably any multi sockets) relies on workarounding this
limitation, by storing interface pointers separately (here in
m_binders). Even if the sc_export looks completely useless in
this situation, it still must be bound.

In this implementation, the sc_export is bound to a dummy interface,
i.e. calls done to it will not be propagated to the target module.
In most case, this is fine, since the get_base_interface() is used
to provide the actual separate interface pointer and is the one
called by the initiator socket.
However, in some other cases like introspection, one may need to
do TLM-2 IMC directly on the sc_export (as demonstrated in the non
regressions).

The following modifications ensures that the export is bound to
the first interface created so that sending a transaction through
the export actually reached the target. This modification is
expected not to break any existing behaviors. In addition, while
the standard does not mention explicitly this corner case, the
general spirit of IEEE 1666-2011/13.2.4.h is that the export
is bound to a channel instance (i.e. TLM-2 interface implementation)
that is provided by the user, not some dummy value.