sc_bitref(_r): return bit_type according to underlying vector type
Instead of always returning `sc_logic` from the implicit conversion
and the bit-wise complement operator, return `bit_type` based on the
current `sc_proxy_traits`.

This enables the expected feature of using a bit reference to an
sc_bv_base in a boolean context:

  sc_bv<7> bv;
  if( bv[2] ) { /* ... */ }

For logic vectors, an explicit comparison with SC_LOGIC_1 or using the
`to_bool` function is still needed.

To enable further consistency, the concatenation references with an
prepended/appended `bool` now use `sc_bv_base` as well.  This automagically
brings the boolean check to concatenations as well.