Add a CMake-based build system.
This initial version of the CMake-based build system has similar
capabilities as the autotools-based build system:

- Configuration of the SystemC sources to use either QuickThreads,
  Pthreads or Fiber for the SystemC coroutines.

- Generation of a set of Makefiles to compile the SystemC libraries
  and accompanying examples.

- Installation of the SystemC library, its header, and documentation
  (optionally also Doxygen-generated source code documentation).

- Execution of the SystemC and TLM examples.

Contrary to the autotools-based build system, it does not fail on
initially unconsidered OS/processor/compiler combinations, as it
configures the sources based on the actually available features on the
target platform. It can not only generate Makefiles, but also Ninja
build scripts, and workspaces for common IDEs such as Visual Studio,
Eclipse, KDevelop, and Xcode.

Installation instructions are provided in INSTALL_USING_CMAKE. A
CMake-specific TODO list is available in the root CMakeLists.txt file.