sc_nbutils.h/get_sign: avoid undefined behaviour (closes #56)
Inverting the value of a signed integral type is undefined behaviour,
if the result can not be represented in the underlying type.

See ISO/IEC 14882:2003(E), clause 5, paragraph 5:

  If during the evaluation of an expression, the result is not mathematically
  defined or not in the range of representable values for its type, the
  behavior is undefined, unless such an expression is a constant expression
  (5.19), in which case the program is ill-formed.

Some compilers do aggressive optimisations and do not care for the overflow
results.  In the case at hand, GCC 4.8 has miscompiled the comparison operators.

To avoid this UB case in the get_sign function, check for the Type's "minimum
value" before applying the unary minus operator.