configure.ac: add configurable installation layout
By default, the files are installed directly to the PREFIX directory root
and the library is installed to PREFIX/lib-<TARGETARCH>, depending on the
current target architecture.  This may be undesired in cases where the
package is meant to be installed in a system-wide location as part of
shared (default) library and include hierarchies (e.g. /usr/local, /usr,
...).

This patch adds an option to follow the the Unix/FHS directory standards:

    --with-unix-layout      use Unix directory layout for installation
                            [default=no]

When "yes", the following (fine-grained) settings will be used:
     --includedir=DIR        C header files [PREFIX/include]
     --libdir=DIR            object code libraries [EPREFIX/lib]
     --docdir=DIR            documentation root [DATAROOTDIR/doc/systemc]

Print a nice summary at the end of the script to summarize the selected
configuration / installation options.