TLM examples: update command-line Makefiles
The command-line Makefiles for the TLM examples  have been updated to
support the 64-bit compilers as well.  Additionally, a joint
"Makefile.rules" has been added, containing the relevant NMake rules for
the projects.  The individual Makefiles only provide the required
variables to set up the build process for each example.

The documentation still needs to be updated.