sc_context.h: avoid reinterpret_cast from 'int' to pointer
In sc_global<T>, the current fixpoint context has been stored as a
pointer to the current process object.  Since this context may be
'NULL', the m_proc member has been initialized by casting '-1' to
a pointer, which is broken.

This commit changes the pointer type for m_proc to a plain 'void *'.
This is sufficient, since only the change of the current context needs
to be detected, whereas the actual value is not of any interest.

The "uninitialized value" can then use an arbitrary address, here the
address of the singleton instance itself.