tracing: removing trace files from simcontext
Previously, trace files have not been removed from the sc_simcontext
upon closing.  This left stale pointers in the trace file list,
eventually leading to segmentation faults, if the simulation continues
afterwards.

This patch adds a "remove_trace_file" counterpart to sc_simcontext.

Note: This is not strictly required by 1666-2011, as it says in [8.1]:

  "A trace file shall not be closed before the final delta cycle
   of simulation."

Accepting less compliant models still seems to be a good idea.