VCD/WIF tracing: unify internal variable naming implementation
Both tracing implementations already contained new functions for
obtaining a new internal variable name without the need for a
temporary variable (as needed by the "old" functions).

This commit streamlines the implementations to use a common function
  "obtain_name()"
to create to new internal variable name.

The old, deprecated functions are dropped from the class definitions
and commented out in the implementation files.

For VCD, this also leads to a bigger number of supported trace variables,
which requires a regolden of the corresponding regression tests.