sc_simcontext: split out sc_status to separate file, add operator<<
The sc_status enum is split from sc_simcontext.h to a separate file
in order to enable a later inclusion from sc_object (for the phase
callbacks).

Secondly, an operator<< for sc_status is added to enable pretty-printing
of sc_status values (and callback masks, i.e. combinations of sc_status
values).