TLM pkg-config/Doxygen: rename from 'tlm-2.0' to 'tlm'
During the integration of the pkg-config support to our ax_systemc.m4
Autoconf macro file, I've noticed that the name 'tlm-2.0.pc' of the
(new) pkg-config file for TLM-2.0 is rather inconvenient (e.g. you can't
easily derive a variable name <pkg>_HOME from the package name, as you
can for SystemC and SCV).

Secondly, the name 'tlm-2.0' is not really in line with the core header
<tlm> and therefore breaks symmetry with <systemc> and <scv.h> as well.

The name of the Doxygen tag file as well as the name of the pkg-config
file (src/tlm_core/tlm.pc.in) are renamed in this patch to remove the
"-2.0" suffix again.

Acked-by: Andy Goodrich <acg@forteds.com>
Acked-by: Torsten Maehne <Torsten.Maehne@lip6.fr>