Print caught sc_report exceptions through sc_report_handler
Triggered by the discussion on the LWG reflector [1] regarding the
relationship of `SC_THROW` and `SC_DISPLAY`, this commit reports the
sc_report exceptions caught in `sc_elab_and_sim` via a call to the
currently installed `sc_report_handler` with an `SC_DISPLAY` action.

This enables a user-defined handling of such messages from within a
custom `sc_report_handler`.

[1] http://www.accellera.org/apps/org/workgroup/lwg/email/archives/201407/msg00001.html