sc_thread_process::suspend_me: elevate assertion to fatal error
When entering the default case of the throw status check (THROWING_NOW),
we have violated the internal invariants if the m_unwinding flag has
been cleared between the entering of the function and the reactivation
of the process.

As this is related to a process restart, the overhead of always checking
for this invariant should not be performance relevant.