sysc/Makefile.am: fix libsystemc_la_LDFLAGS
The path '-L$(libarchdir)', pointing to the library installation
directory might not exist yet and should not include any relevant
libraries for linking the current SystemC library.  If this directory
is necessary for linking on a specific platform, the user can pass
it via the environment variable LDFLAGS to the build system.

Secondly, EXTRA_LDFLAGS is added to this overridden variable.