configure.ac: suppress unused command-line argument warnings of clang
Clang complains about command-line arguments that are not used by the
different invoked driver programs.  In our setup, this currently affects
some assembler and linker arguments, that are forwarded although not
strictly needed.

Ignore these harmless warnings for the assembler and linker calls through
clang, if the default assembler and linker is used.  Closes #79.