sc_export: move binding check to 'elaboration_done'
As discussed in [1], the LRM currently requires an error at the end of
elaboration, in case an export is not bound (5.13.8).

Instead of checking the binding of an export already before
'before_end_of_elaboration', postpone the check until before
'end_of_elaboration' to lift the overly strict check.

For consistency with port bindings, use 'SC_ID_COMPLETE_BINDING_'
for the error message.

[1] http://www.accellera.org/apps/org/workgroup/lwg/email/archives/201408/msg00006.html