sc_vector_iter: use const (direct) iterator for mixed expressions
The sc_vector iterators (see 1666-2011, 8.5.8) are meant to be
transparently usable with respect to conversions to const and
from "member iterators" (sc_assemble_vector) to the direct element
iterators.

For the comparison (and difference) operations, this implies that
mixed const-ness iterator expressions should be supported (as it
would the case for plain pointers):
  sc_vector<T>::const_iterator citr = vec.begin(); // ok (conversion)
  sc_assert( citr == vec.begin() );                // ok
  sc_assert( vec.begin() == citr );                // fails
  sc_assert( vec.size() == vec.end() - citr );     // fails (difference)

The latter expressions currently fail, as the iterator operators
used to require the same iterator as second parameter.

A similar limitation exists for the mixed comparison with "member
iterators", again with the "special" iterators on the left-hand side:
  sc_assert( vec.end() == sc_assemble_vector(vec, &T::m).end() ); // ok
  sc_assert( sc_assemble_vector(vec, &T::m).end() == vec.end() ); // fails

This commit lifts the restriction by always expecting a const_iterator
with the direct_access policy (vector element).  As all sc_vector<T>
(member) iterators can be converted to this type, all combinations
should work as expected.

See OSCI-WG/systemc-regressions@316d23fa for a related test case, which
passes only after this patch is applied.