segmentation fault in circular_buffer:clear()
In the implementation-defined class tlm::circular_buffer, the clear()
function may call the destructor of previously destroyed elements.
This can happen in case of a read index m_ri != 0.

Properly adjust the memory locations to be freed based on the current
circular window.

This is related to but different from #35.