intel.hpp (Boost): add disambiguating namespace
In the Boost config for the Intel compiler, a static assertion check for
(intrinsic) wchar support is included.  This check is based on the
definition of a helper template named `assert_(no)instrinsic_wchar_t`.

Being defined in the global namespace, this leads to conflicts with
external Boost installations (when using the Intel compiler).

This patch moves the helper to the 'sc_boost' namespace, where the other
Boost symbols are located in the embedded sources in SystemC already.