Ensure that WAIT_UNTIL can handle nested WAIT calls.
With the previous implementation a nested WAIT call in WAIT_UNTIL
would set the file and lineno information (correctly), but once the
thread is again waiting in the outer WAIT_UNTIL the information would
be wrong, since it was not updated when re-entering the outer wait.

Fixes test added in 8211919ba1cc47a080cfe59d422055755f199967.