sc_vector.h: hide warning about missing std::vector<void*> export
As sc_vector_base is using an std::vector<void*> for storage internally,
MSVC complains about a potentially missing DLL export for this type.

Usually, we would export such instantiations explicitly, but as the value
type is void* here, we don't do this here to avoid conflicts with other
sources using/exporting this type.