Add API annotations to build DLL on Windows
This commit adds the first set of annotations to build SystemC as DLL
on Windows.  For that purpose, the following macros are introduced in
sysc/kernel/sc_cmnhdr.h:

 * SC_WIN_DLL - activate DLL annotations on Windows
   This needs to be set during both, library and application build.

 * SC_BUILD   - indicate the build of the SystemC library itself

 * SC_API     - when SC_WIN_DLL is set, mark a symbol for
   - DLL export (during library build)
   - DLL import (during application build)

The SC_API macro currently expands to nothing on non-Windows platforms
or if a static library is built (i.e. SC_WIN_DLL is unset).