Add template instantiations to library (and export from DLL)
Some template classes are instantiated within the SystemC library
itself.  Especially when building a SystemC DLL, these template
instantiations need to be added to (and exported from) the library.

To achieve that, such template instantiations are marked as "extern",
if SC_BUILD is not set (which is part of C++11, but supported by MSVC
and GCC (compatible) compilers).  The internal macro SC_API_TEMPLATE_
is used to handle this selection.

In addition to SystemC template class specializations, some std::vector<>
instantiations are pre-built as well (for SystemC-specific pointer/value
types).  The helper macro SC_API_VECTOR_ handles these exports.