sc_main_main.cpp: include sc_ver.h to include version check
To enable building a Windows DLL, SystemC needs to be split into
a "core" library (containing the kernel itself) and the "full"
library, depending on the core library and providing the additional

 * main
 * sc_elab_and_sim

functions, which can only be linked statically.

In order to leverage the API compatibility check provided by sc_ver.h,
the sc_main_main.cpp includes this symbol as well, ensuring link-time
compatibility between the core library (DLL) and the static library.