configure.ac: fix autoconf run on Autoconf 2.69
The workaround for older Autoconf versions introduced in #119
led to a (false-positive) error on Autoconf 2.69:

configure.ac:64: error: possibly undefined macro: AS_VAR_APPEND
      If this token and others are legitimate, please use m4_pattern_allow.
      See the Autoconf documentation.

This seems to be triggered by a using a '#' instead of the 'dnl' marker
at the beginning of of the leading comment.  Changing it avoids the error,
while keeping the support for Autoconf 2.62.