utils/sc_string_view.h: add non-owning, constant reference to string
This file adds the class/typedef sc_core::sc_string_view to the
utils directory of the SystemC kernel.  This class is meant to follow
the latest C++ Library Fundamentals TS [1] for providing a non-owining,
constant reference to character strings.

With this helper, many APIs dealing with strings can be implemented more
efficiently by reducing the number of temporary string copies.

[1]: http://open-std.org/jtc1/sc22/wg21/docs/papers/2015/n4480.html#string.view