Fix G++ 6.1 warning regarding the use of deprecated std::auto_ptr<T>.
Origin of the warning is the partial specialization of
sc_boost::get_pointer<T> for std::auto_ptr<T>. This patch backports the
local disabling of the compiler warning, which is present in Boost 1.59.
Be aware that boost::get_pointer<T> has been specialized in Boost 1.59
also for the other smart pointers introduced with C++'11
std::unique_ptr<T> and std::shared_ptr<T>. These additional
specializations are not introduced with this commit.

Anyway, it should be considered to upgrade the Boost version shipped
with SystemC to consider fully the features (and bugs) of C++'>=11 and
modern compiler versions.