Don't use POSIX function strdup() in sc_core::sc_elab_and_sim().
The strdup() function is not part of the ISO C++'03 standard. As that
standard reserves function names beginning with "str", MSVC does issue
a warning when using strdup() and offers _strdup as a replacement.
However, that replacement is not part of the C++ standard libraries
shipped with GCC and Clang on Linux and OS X. Therefore, the single use
of strdup() in the SystemC implementation is replaced with an equivalent
call sequence to malloc(), strlen(), and strcpy().