Add missing includes after removal of sc_iostream.h
After the removal of sc_iostream.h in osci-wg/systemc#165, the <cstring>
header is no longer implicitly included via <systemc> (or other parts
in the library).

This break compilation on some platforms due to missing declaration
of e.g. std::memcpy or std::strcmp.

This patch adds the required includes to the affected files (sometimes
cleaning the include order along the way) and adds an opt-in switch to
readd <cstring> inclusion to <systemc> to reduce migration cost for
client code.