Optional inclusion of <sstream> from <systemc>
With the removal of sc_iostream.h in #165, the list of implicit
includes from <systemc> has been reduced.  The SystemC examples
and tests only required <cstring>, see #170.

In osci-wg/uvm-systemc#96, it was noted that <sstream> could
also be missing in client code.

This patch changes the flag `SC_INCLUDE_CSTRING` to
`SC_INCLUDE_CSTRING_SSTREAM` and adds `<sstream>` to the
includes guarded by the switch.