misc/unit/control/wait: keep process alive
The process `dut.thread` ends during the execution of the test,
leading to an immediate release of its resources after the change
in osci-wg/systemc#183.  As a result, the test accesses invalid memory
after the final sc_start call.

This is avoided by keeping the process alive through a corresponding
`sc_process_handle` object.