tlm_utils::instance_specific_extension: move to library
This patch moves necessary infrastructure parts of the instance-specific
extension classes to the pre-built library.

The required TLM generic payload extension is still kept in the header
in order to avoids its registration with the main extension mechanism
in case no instance-specific extensions are used in the model.

Extension type registration is tracked similarly as for the generic
payload (see cd085d8b).

By changing the accessor internals slightly, arbitrary payload types
(providing the required extension functions) are now supported.

Change-Id: I946507dd95d39d7079ab6796420263bdd80b5c70