Introduce alias SystemC::systemc to target library systemc.
This will be the name of the exported target, which future CMake-based
build systems of applications may use to link against the
SystemC library. For the moment, only the the SystemC and TLM examples
make already use of SystemC::systemc. Using this approach is more robust
than the previous approach that used a CMake variable to specify the
library to link to. In case of errors, CMake will issue a more sensible
error message for the target alias/imported target than for a variable,
which may be undefined or have a typo in its name.