Fix "make check" errors due to invalid shell variable names
As reported in the user forums, "make check" fails on some systems
due to test paths starting with a digit (e.g. 2.1/dpipe).  This error
is caused by the restriction of some shells to not accept variable
names starting with a digit at all.

As a solution, this patch generates/substitutes the filter value
in-place from the Makefile, instead of synthesizing the "FILTER"
variable inside the script.

This also removes the need to export the filter variable to the
environment.