sc_process_table: fixes and deduplication
The Coverity static analysis tool found dead code in the
sc_process_table::remove functions, where the prior_p
pointer was never updated:

    prior_p = 0;
    for ( now_p = m_thread_q; now_p; now_p = now_p->next_exist() )
    {
        if ( now_p == handle_ )
        {
           if ( prior_p )  //<-- always NULL
                prior_p->set_next_exist( now_p->next_exist() );
           else
                m_thread_q = now_p->next_exist();
            return handle_;
        }
    }

Therefore, even if the found handle is not the first one, the first
entry will be dropped from the table.

This commit fixes this bug and adds a template class to avoid
the duplicated logic between the method and thread queue handling.

Reported-by: Andreas Hedstrom <andreas.hedstrom@intel.com>