tlm_array: drop unused default value
tlm_array used to store a default value explicitly, even though it is
not used externally and there are very little uses for a non-NULL
default value of a pointer type.

As this class is implementation-defined anyway, we can reduce the size
of tlm_array (and therefore of tlm_generic_payload) by the size of one
pointer when dropping this member.

No functional changes.