Alter the simple_async example to be suitable as a regression test
The example now works both with and without c++11, though without it simply
uses the async event from within systemc.
A track is kept to ensure the event has fired before the end of simulation.

To enable the regression test, a golden log has been added (and a vcxproj file)

Signed-off-by: BURTON Mark <mark.burton@greensocs.com>