simple_target_socket: only spawn b2nb_thread, if needed
The b2nb_thread process in the simple_target_socket(_tagged)
is only ever utilized in case no blocking, but only a
non-blocking transport callback is registered.

Defer the spawning of this thread to the start_of_simulation
phase and only spawn the thread in this scenario.

Secondly, switch to static sensitivity and defer the
initialization of the process until the first notification.

No functional changes, apart from the fact the user classes
deriving from simple_target_socket(_tagged) need to call the
base classes's start_of_simulation callback, if overridden.