Document pthread dependency of async updates (closes #260)
Linking against a static SystemC library with asynchronous update
support enabled may require an explicit link against the pthread
library as well (due to use of mutexes and semaphores).

For dynamic libraries, this is not needed as the lib itself is
already properly linked in this case.

Add notes to the INSTALL file to document this case.