sc_event: remove dynamic sensitivity in destructor
In case an event is destroyed while some processes are still
(dynamically) sensitive to it, a use-after-free can happen
during reset or destruction of such processes, which will try
to remove itself from the dynamic sensitivity of an already
destroyed event.

To avoid this, clear all dynamically sensitive processes of
an event during its destruction.

Note: Adding a warning is not justified, as this can regularly
      happen during teardown of the simulation.

Caught by the AddressSanitizer [1].

[1]: https://github.com/google/sanitizers/wiki/AddressSanitizer