sc_port_base: free bind info in destructor
In case of (binding) errors during elaboration, the m_bind_info member
of ports is not always cleaned up correctly.  This patch moves the
functionality to cleanup the binding storage into a separate function
`free_binding`, which is then called from the destructor as well.

Additionally, the vectors are fully emptied (including the release
of their allocated buffer) by using the swap() idiom.