sc_thread_process: remove dynamic sensitivity before dying
When a running thread process is killed during simulation,
an exception is thrown in the stack of the process.  If
the thread is currently waiting on an sc_event_list stored
on its own stack, this event list will be destroyed
during stack unwinding.

As event lists enforce during their destruction that no
process is still waiting on them, this check fails in the
above case.

To avoid this failing check, this commit removes any
dynamic sensitivity from a thread before throwing the
exception.

Static sensitivity or method processes are not affected
by this corner case.