Require semicolon after SC_METHOD/SC_(C)THREAD/SC_NEW
It is typically a recommended practice to require a semicolon
after "statement-like" macros in C++.  In SystemC, the process
macros previously did not require this.

This commit changes the macro definitions to exclude the semicolon.

For backwards-compatibility, a new SC_ALLOW_MACROS_WITHOUT_SEMICOLON
switch is added, allowing to omit the semicolon for the macros

 - SC_(C)THREAD
 - SC_METHOD
 - SC_NEW (non-standard)