remove shifts of negative values, and negative shifts
There were a number of compile errors for negative shifts (on value and
argument). Changed hardcoded -1 and -2 to use numerical limits to populate bit
patterns, and also to get number of bits. This is parameterized, using the
traits of sc_digit. It is not done fully throughout the code, this to bre done
later.

This checkin also remove trailing whitespace, so more diffs than needed.