Remove call to sc_abort() from sc_assertion_failed()
sc_assertion_failed() calls the default report handler which in turn
will call sc_abort() if the SC_ABORT action is set.

Currently, it is not possible to avoid calling sc_abort() even with
setting the SC_FATAL report actions to SC_DO_NOTHING.

When using SystemC with Verilator, and abort causes no trace to be
written and therefore makes sc_assert() unusable for runtime checks.