Don't use `force_align_arg_pointer` on non-x86 platforms
Currently, `SC_ALIGNED_STACK_` is defined to `__attribute__((force_align_arg_pointer))` on any platform that's not x86_64, but this attribute only exists for the gcc/clang x86 backends (https://gcc.gnu.org/onlinedocs/gcc/x86-Function-Attributes.html). With this change, the attribute is only used on i386, and not on other platforms (e.g. aarch64) where it would be unexpected.

Signed-off-by: Lexi Bromfield <lexinadia@google.com>