Fix undefined behaviour when assigning a negative value to an sc_int
The previous code here attempted to perform a sign extension by shifting
left and then right to fill the top bits with sign bit.  If the input is
negative then this applies a left shift to a negative value, which is
undefined behaviour in C++17 and the error is caught by ubsan.  Fix this
by explicitly picking out the sign bit and then filling the top bits
with that value.

Signed-off-by: Mark Thompson <mark.thompson@amd.com>