Enable shared library build without global flag
SystemC's CMakeLists.txt used to rely on the global BUILD_SHARED_LIBS
flag to enable building SystemC as a shared library. This also forces
all other libraries in the project to be build as shared libraries. This
commit enables the user to select only SystemC to be build as a shared
library without forcing this on all other libraries in the cmake
project.

Signed-off-by: Lukas Jünger <lukas@mwa.re>