Add SYSTEMC_UNITY_BUILD CMake option and require CMake 3.16+
This patch introduces a new CMake option -DSYSTEMC_UNITY_BUILD=ON to enable
unity builds for the SystemC library, which may improve build performance
in some environments.

Enabling this option compiles the sources as a single compilation unit.
The option is OFF by default and does not affect standard builds.

The minimum required CMake version is raised to 3.16 to support this feature.

Signed-off-by: Ruben Brandhofer <ruben@mwa.re>