Windows: Refactor DLL creation on Cygwin/MinGW env
Windows DLLs can't contain unresolved symbols. Therefore systemc,
when built with MSYS2, is split in two libraries: libsystemc-<version>.dll
without unresolved symbols and libsystemc.a with unresolved symbols.

Currently, cmake target systemc (exposing libsystemc.a) is aliased as SystemC::systemc

This leads to a situation where projects that depend SystemC::systemc
would link only against the systemc.a and drop the dll.

This commit introduces the interface target library `systemc_if`,
which links against systemc dll and static and is aliased as SystemC::systemc.

To insure backward compatiblity, this mechanism is enabled only
when systemc is built as a shared library

This commits reworks the sections that merges
libsystemc.a and libsystemc-<version>.dll.a